Generalized ResponseSquaredL2Difference[Side] to work with any target uniform value
Before, you could specify 'ZERO' as the target field name,
causing the evaluator to compute the norm of the source field.
Now, one can specify 'Target Value' (and *not* specify any
target field name) and compute the squared L2 norm of the
different between the source field and the specified value.