Changed implementation style for the build type static setting
It's now wrapped inside the Albany::Session struct, and can be reset.
This helps if a customer app wants to run albany twice in the same
build, with two different build types (e.g., compare the outputs, or
one of the two runs requires something peculiar to a build type).

The Session struct is also easy to extend, to add centralized handling
of other things, like MPI or Kokkos.