Bug fix in STK discretization, plus added option
The ss discretizations were created using the parameter list of the
volume discretization, rather than their own sublists.

Also, added possibility to add all states to the NodalDOFsStructContainer.
It's not really a big deal: if there was already another dof on that part
with the same dimension, there's no additional DOFStruct built.