Bug fix in the creation of the Epetra FECrs matrix.
When creating an Epetra_FECrsMatrix, we require a column map (the overlapped
domain map). This is because if Epetra creates it during the fill complete
call, it will have gids ordered in a way different from the one we use in
our overlapped maps in the discretization. This would make the GID<->LID
mapping inside the FE matrix inconsistent with the one we get from the
vector spaces in the discretization.