Adding a LayeredMeshNumbering<GO> object in the mesh struct.
Ideally, we should switch to only use this, since the switch to FECrs matrix
in Tpetra requires all owned nodes to come first in the map, which causes
nodes to have local ordering inside the map that does not satisfy the
layered mesh numbering scheme anymore.

Instead, one should do the id*stride+offset arithmetic on global indices.
If local ids are needed, one should get the GlobalLocalIndexer (corresponding
to the vector space of the GID queried) from the discretization